+++ ts/packages/marginfi-cli/src/decode-event.ts
@@ -1,4 +1,3 @@
-import { Decimal } from "@mrgnlabs/marginfi-client";
 import { OptionValues } from "commander";
 import { getClientFromOptions } from "./common";