+++ ts/packages/marginfi-client/src/types.ts
@@ -13,12 +13,12 @@ export type UiAmount = BigNumber | number | I80F48 | string;
 
 export enum UtpIndex {
   Mango = 0,
-  Zo = 1,
+  ZO = 1,
 }
 
 export const UTP_NAME = {
   [UtpIndex.Mango]: "Mango",
-  [UtpIndex.Zo]: "01",
+  [UtpIndex.ZO]: "01",
 };
 
 /**
@@ -27,15 +27,6 @@ export const UTP_NAME = {
 export enum Environment {
   DEVNET = "devnet",
   MAINNET = "mainnet",
-}
-
-/**
- * Marginfi bank vault type
- */
-export enum BankVaultType {
-  LiquidityVault,
-  InsuranceVault,
-  FeeVault,
 }
 
 export interface MarginfiDedicatedConfig {
@@ -142,36 +133,32 @@ export interface MarginfiGroupData {
 export interface MarginfiAccountData {
   authority: PublicKey;
   marginfiGroup: PublicKey;
-  depositRecord: DecimalData;
-  borrowRecord: DecimalData;
+  depositRecord: WrappedI8048F;
+  borrowRecord: WrappedI8048F;
   activeUtps: boolean[];
   utpAccountConfig: UTPAccountConfig[];
   reservedSpace: BN[];
 }
 
 export interface BankData {
-  scalingFactorC: DecimalData;
-  fixedFee: DecimalData;
-  interestFee: DecimalData;
-  depositAccumulator: DecimalData;
-  borrowAccumulator: DecimalData;
+  scalingFactorC: WrappedI8048F;
+  fixedFee: WrappedI8048F;
+  interestFee: WrappedI8048F;
+  depositAccumulator: WrappedI8048F;
+  borrowAccumulator: WrappedI8048F;
   lastUpdate: BN;
-  nativeDepositBalance: DecimalData;
-  nativeBorrowBalance: DecimalData;
+  totalDepositsRecord: WrappedI8048F;
+  totalBorrowsRecord: WrappedI8048F;
   mint: PublicKey;
   vault: PublicKey;
-  bankAuthorityBump: number;
+  bankAutorityBump: number;
   insuranceVault: PublicKey;
-  insuranceVaultAuthorityBump: number;
+  insuranceVaultAutorityBump: number;
   feeVault: PublicKey;
-  feeVaultAuthorityBump: number;
-  initMarginRatio: DecimalData;
-  maintMarginRatio: DecimalData;
+  feeVaultAutorityBump: number;
+  initMarginRatio: WrappedI8048F;
+  maintMarginRatio: WrappedI8048F;
 }
-
-export interface DecimalData {
-  flags: number;
-  hi: number;
-  lo: number;
-  mid: number;
+export interface WrappedI8048F {
+  bits: BN;
 }